+++ src/main/java/com/mock/interview/interviewquestion/presentation/web/list/QuestionListWebController.java
@@ -1,11 +1,11 @@
-package com.mock.interview.interviewquestion.presentation.web;
+package com.mock.interview.interviewquestion.presentation.web.list;
 
 import com.mock.interview.interview.infra.InterviewQuestionRepositoryForView;
 import com.mock.interview.interviewquestion.infra.QuestionRepositoryForListView;
-import com.mock.interview.interviewquestion.infra.QuestionRepositoryForView;
 import com.mock.interview.interviewquestion.presentation.dto.QuestionOverview;
 import com.mock.interview.interviewquestion.presentation.dto.QuestionSearchCond;
 import com.mock.interview.interviewquestion.presentation.dto.QuestionSearchOptionsDto;
+import com.mock.interview.interviewquestion.presentation.web.QuestionPageInitializer;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -18,8 +18,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 @Controller
 @RequiredArgsConstructor
-public class QuestionSearchWebController {
-    private final QuestionRepositoryForView questionRepositoryForView;
+public class QuestionListWebController {
     private final QuestionRepositoryForListView questionRepositoryForListView;
     private final InterviewQuestionRepositoryForView interviewQuestionRepositoryForView;
 
@@ -46,36 +45,6 @@ public class QuestionSearchWebController {
         Page<QuestionOverview> overviewPage = interviewQuestionRepositoryForView.findOverviewList(interviewId, pageable);
         QuestionPageInitializer.initListPage(model, overviewPage, new QuestionSearchCond(/* 임시코드 */), request);
         return "interview/question";
-    }
-
-    @GetMapping("question/category/{categoryId}")
-    public String categoryQuestionListPage(
-            @PathVariable(name = "categoryId") long categoryId,
-            QuestionSearchCond searchCond,
-            Model model, @PageableDefault Pageable pageable,
-            HttpServletRequest request
-    ) {
-        QuestionSearchOptionsDto searchOptions = QuestionSearchOptionsDto.builder()
-                .categoryIdCond(categoryId)
-                .searchCond(searchCond).build();
-        Page<QuestionOverview> overviewPage = questionRepositoryForListView.findOverviewList(searchOptions, pageable);
-        QuestionPageInitializer.initListPage(model, overviewPage, searchCond, request);
-        return "question/list";
-    }
-
-    @GetMapping("question/position/{positionId}")
-    public String positionQuestionListPage(
-            @PathVariable(name = "positionId") long positionId,
-            QuestionSearchCond searchCond,
-            Model model, @PageableDefault Pageable pageable,
-            HttpServletRequest request
-    ) {
-        QuestionSearchOptionsDto searchOptions = QuestionSearchOptionsDto.builder()
-                .positionIdCond(positionId)
-                .searchCond(searchCond).build();
-        Page<QuestionOverview> overviewPage = questionRepositoryForListView.findOverviewList(searchOptions, pageable);
-        QuestionPageInitializer.initListPage(model, overviewPage, searchCond, request);
-        return "question/list";
     }
 
     @GetMapping("question/{questionId}/child")
